+++ packages/moonstone/ProgressBar/tests/ProgressBar-specs.js
@@ -7,7 +7,7 @@ describe('ProgressBar Specs', () => {
 	it('Should have width of 50%', () => {
 		const progressBar = mount(
 			<ProgressBar
-				progress={50}
+				progress={0.5}
 			/>
 		);
 
@@ -17,7 +17,7 @@ describe('ProgressBar Specs', () => {
 	it('Should have background width of 75%', () => {
 		const progressBar = mount(
 			<ProgressBar
-				backgroundProgress={75}
+				backgroundProgress={0.75}
 			/>
 		);
 
@@ -27,7 +27,7 @@ describe('ProgressBar Specs', () => {
 	it('Should have height of 50%', () => {
 		const progressBar = mount(
 			<ProgressBar
-				progress={50}
+				progress={0.50}
 				vertical
 			/>
 		);
@@ -38,26 +38,12 @@ describe('ProgressBar Specs', () => {
 	it('Should have background height of 50%', () => {
 		const progressBar = mount(
 			<ProgressBar
-				progress={50}
-				backgroundProgress={75}
+				progress={0.50}
+				backgroundProgress={0.75}
 				vertical
 			/>
 		);
 
 		expect(progressBar.find(`.${css.load}`).prop('style').height).to.equal('75%');
-	});
-
-	it('Should have progress height of 25%', () => {
-		const progressBar = mount(
-			<ProgressBar
-				progress={50}
-				backgroundProgress={75}
-				min={0}
-				max={200}
-				vertical
-			/>
-		);
-
-		expect(progressBar.find(`.${css.fill}`).prop('style').height).to.equal('25%');
 	});
 });